+++ fuel-tx/src/transaction/consensus_parameters/gas/default_gas_costs.rs
@@ -19,7 +19,6 @@ pub fn default_gas_costs() -> GasCostsValues {
         divi: 1,
         eck1: 951,
         ecr1: 3000,
-        ed19: 3000,
         eq: 1,
         exp: 1,
         expi: 1,
@@ -100,6 +99,10 @@ pub fn default_gas_costs() -> GasCostsValues {
         },
         cfei: DependentCost::LightOperation {
             base: 2,
+            units_per_gas: 214,
+        },
+        ed19: DependentCost::LightOperation {
+            base: 3000,
             units_per_gas: 214,
         },
         k256: DependentCost::LightOperation {